Create first version of etch-a-sketch project
Create html file with boilerplate skeleton, a container div, and a button.
Then create css file with basic styling for the container and button including
typography, colors, and sizes. Finally, create a javascript file for scripting
the etch-a-sketch functionality within the container.

Use DOM manipulation and event handling to create the container of the correct
size. Within css file, utilize the css grid properties within the container
element to organize the children divs appropriately.

Create event handlers for container's children divs so they are set a class name
when hovered over by the user's mouse. This applied class will change the
background color of that div via the css stylesheet.

Finally, create an event handler for the button element which will
prompt the user for their desired sketchbook/grid size and reset the grid
to that size.